Edit Dream model
