+++ JS-Sql/database/models/index.js
@@ -2,7 +2,6 @@
 
 const fs = require('fs');
 const path = require('path');
-const Sequelize = require('sequelize');
 const basename = path.basename(__filename);
 const sequelize = require('../connection');
 const db = {};
@@ -27,6 +26,5 @@ Object.keys(db).forEach((modelName) => {
 });
 
 db.sequelize = sequelize;
-db.Sequelize = Sequelize;
 
 module.exports = db;
